# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Migration Bench — Min v0 API
  version: "0.0.0"
  description: |
    Minimal API surface for migration benchmarking.

    Invariants:
    - All `/api/*` routes (except `/api/health`) require `x-ops-user-id`.
    - 400/403 responses use the `VALIDATION_ERROR` envelope.
    - 404 responses are `{ "message": "Not found" }`.
    - `GET /api/activity?forceError=true` returns HTTP 500 with an `INTERNAL_ERROR` envelope.
servers:
  - url: http://localhost:3001
paths:
  /api/health:
    get:
      summary: Health check
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  ok: { type: boolean }
                required: [ok]

  /api/users:
    get:
      summary: List users
      parameters:
        - $ref: "#/components/parameters/XOpsUserId"
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/User" }
        "403":
          $ref: "#/components/responses/ValidationError403"

  /api/incidents:
    get:
      summary: List incidents
      parameters:
        - $ref: "#/components/parameters/XOpsUserId"
        - name: status
          in: query
          required: false
          schema: { $ref: "#/components/schemas/IncidentStatus" }
        - name: severity
          in: query
          required: false
          schema: { $ref: "#/components/schemas/IncidentSeverity" }
        - name: q
          in: query
          required: false
          schema: { type: string }
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [createdAt_desc, createdAt_asc, severity_asc, severity_desc]
        - name: breached
          in: query
          required: false
          description: String boolean ("true"/"false") per server implementation.
          schema:
            type: string
            enum: ["true", "false", ""]
      responses:
        "200":
          description: Incident list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Incident" }
        "403":
          $ref: "#/components/responses/ValidationError403"
    post:
      summary: Create incident (responder/admin)
      parameters:
        - $ref: "#/components/parameters/XOpsUserId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/IncidentCreate" }
      responses:
        "201":
          description: Created incident
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Incident" }
        "400":
          $ref: "#/components/responses/ValidationError400"
        "403":
          $ref: "#/components/responses/ValidationError403"

  /api/incidents/{id}:
    get:
      summary: Get incident by id
      parameters:
        - $ref: "#/components/parameters/XOpsUserId"
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Incident
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Incident" }
        "403":
          $ref: "#/components/responses/ValidationError403"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      summary: Patch incident (responder/admin)
      parameters:
        - $ref: "#/components/parameters/XOpsUserId"
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/IncidentPatch" }
      responses:
        "200":
          description: Updated incident
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Incident" }
        "400":
          $ref: "#/components/responses/ValidationError400"
        "403":
          $ref: "#/components/responses/ValidationError403"
        "404":
          $ref: "#/components/responses/NotFound"

  /api/incidents/{id}/reopen:
    post:
      summary: Reopen incident (resolved → investigating) (responder/admin)
      parameters:
        - $ref: "#/components/parameters/XOpsUserId"
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Reopened incident
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Incident" }
        "400":
          $ref: "#/components/responses/ValidationError400"
        "403":
          $ref: "#/components/responses/ValidationError403"
        "404":
          $ref: "#/components/responses/NotFound"

  /api/activity:
    get:
      summary: List activity (newest-first)
      parameters:
        - $ref: "#/components/parameters/XOpsUserId"
        - name: forceError
          in: query
          required: false
          schema: { type: string }
          description: When "true", returns HTTP 500 deterministically.
      responses:
        "200":
          description: Activity list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Activity" }
        "403":
          $ref: "#/components/responses/ValidationError403"
        "500":
          description: Deterministic failure mode for testing
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InternalError" }

components:
  parameters:
    XOpsUserId:
      name: x-ops-user-id
      in: header
      required: true
      schema: { type: string }
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
  responses:
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/NotFound" }
    ValidationError400:
      description: Validation error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ValidationError" }
    ValidationError403:
      description: Forbidden / missing user context
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ValidationError" }
  schemas:
    UserRole:
      type: string
      enum: [admin, responder, viewer]

    IncidentStatus:
      type: string
      enum: [triggered, acknowledged, investigating, mitigated, resolved]

    IncidentSeverity:
      type: string
      enum: [S1, S2, S3, S4]

    User:
      type: object
      additionalProperties: false
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string }
        role: { $ref: "#/components/schemas/UserRole" }
        team: { type: string }
        onCall: { type: boolean }
      required: [id, name, email, role, team, onCall]

    Incident:
      type: object
      additionalProperties: false
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string }
        status: { $ref: "#/components/schemas/IncidentStatus" }
        severity: { $ref: "#/components/schemas/IncidentSeverity" }
        commanderId:
          oneOf: [{ type: string }, { type: "null" }]
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        acknowledgedAt:
          oneOf: [{ type: string, format: date-time }, { type: "null" }]
        resolvedAt:
          oneOf: [{ type: string, format: date-time }, { type: "null" }]
        tags:
          type: array
          items: { type: string }
      required: [id, title, description, status, severity, commanderId, createdAt, updatedAt, acknowledgedAt, resolvedAt, tags]

    IncidentCreate:
      type: object
      additionalProperties: false
      properties:
        title: { type: string, minLength: 4 }
        description: { type: string, minLength: 1 }
        severity: { $ref: "#/components/schemas/IncidentSeverity" }
        tags:
          type: array
          items: { type: string }
      required: [title, description, severity]

    IncidentPatch:
      type: object
      additionalProperties: false
      properties:
        title: { type: string, minLength: 4 }
        description: { type: string, minLength: 1 }
        severity: { $ref: "#/components/schemas/IncidentSeverity" }
        tags:
          type: array
          items: { type: string }
        commanderId:
          oneOf: [{ type: string }, { type: "null" }]
        status: { $ref: "#/components/schemas/IncidentStatus" }

    Activity:
      type: object
      additionalProperties: false
      properties:
        id: { type: string }
        type:
          type: string
          enum: [create, update, delete, reopen]
        entityType:
          type: string
          enum: [incident]
        entityId: { type: string }
        message: { type: string }
        createdAt: { type: string, format: date-time }
      required: [id, type, entityType, entityId, message, createdAt]

    ValidationError:
      type: object
      additionalProperties: false
      properties:
        error:
          type: object
          additionalProperties: false
          properties:
            code:
              type: string
              enum: [VALIDATION_ERROR]
            message: { type: string }
            fields:
              type: object
              additionalProperties: { type: string }
          required: [code, message, fields]
      required: [error]

    NotFound:
      type: object
      additionalProperties: false
      properties:
        message:
          type: string
          enum: ["Not found"]
      required: [message]

    InternalError:
      type: object
      additionalProperties: false
      properties:
        error:
          type: object
          additionalProperties: false
          properties:
            code:
              type: string
              enum: [INTERNAL_ERROR]
            message: { type: string }
          required: [code, message]
      required: [error]